% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm.R
\name{svm}
\alias{svm}
\title{Solving Regularized Support Vector Machine}
\usage{
svm(x, y, C = 1, max_iter = 1000, tol = 1e-05, shrink = TRUE, verbose = 0)
}
\arguments{
\item{x}{The data matrix \eqn{X=(x_1,\ldots,x_n)^T} of size
\eqn{n\times d}, representing \eqn{n} observations
and \eqn{d} features.}

\item{y}{The length-\eqn{n} response vector.}

\item{C}{The cost parameter.}

\item{max_iter}{Maximum number of iterations.}

\item{tol}{Tolerance parameter for convergence test.}

\item{shrink}{Whether to use the shrinkage algorithm.}

\item{verbose}{Level of verbosity.}
}
\value{
A list of the following components:
\item{beta}{Optimized value of the \eqn{\beta} vector.}
\item{xi,Lambda,Gamma}{Values of dual variables.}
\item{niter}{Number of iterations used.}
\item{dual_objfns}{Dual objective function values during the optimization process.}
}
\description{
This function solves the regularized support vector machine
of the following form:
\deqn{
\min_{\beta}\ \frac{C}{n}\sum_{i=1}^n \max(1-y_i x_i^T\beta,0) +
\frac{1}{2}\Vert\beta\Vert_2^2
}
where \eqn{\beta\in\mathbb{R}^d} is a length-\eqn{d} vector,
\eqn{x_i\in\mathbb{R}^d} is the feature vector for the \eqn{i}-th observation,
\eqn{y_i\in\{-1,1\}} is a binary label,
and \eqn{C} is the cost parameter.
}
\author{
Yixuan Qiu \url{https://statr.me}

        Ben Dai \url{https://bendai.org}
}
